Module 11 Python and mysql/requirements.txt
mysql-connector-python==8.3.0
protobuf==4.25.2
setuptools==69.0.3
wheel==0.42.0
